Refactor order and payment processing components
- Updated IActiveOrderRepository to extend IOrderRepository, simplifying the interface.
- Changed charge method in StubPaymentGateway to accept BigDecimal instead of double for better precision.
- Removed unused imports and cleaned up test files for CompanyLifecycleServiceTest, EventServiceCancelEventTest, and EventServiceCreateEventTest.
- Deleted MoneyTest and refactored related tests in OrderServiceTest to ensure coverage of payment and order functionalities.
- Enhanced OrderServiceTest with additional scenarios for ticket purchasing, including handling of discounts and member orders.
- Updated PolicyFrameworkTest to remove unnecessary dependencies and streamline policy application tests.
- Adjusted StubGatewaysTest to align with new PaymentDetails structure and ensure proper testing of payment gateway functionality.

+++ src/main/java/com/ticketing/infrastructure/gateway/StubPaymentGateway.java
@@ -6,6 +6,7 @@ import com.ticketing.domain.gateway.PaymentDetails;
 import com.ticketing.domain.gateway.PaymentResult;
 import com.ticketing.domain.gateway.RefundResult;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class StubPaymentGateway implements IPaymentGateway {
@@ -17,7 +18,7 @@ public class StubPaymentGateway implements IPaymentGateway {
     }
 
     @Override
-    public PaymentResult charge(double amount, PaymentDetails details) {
+    public PaymentResult charge(BigDecimal amount, PaymentDetails details) {
         if (shouldFail) {
             return PaymentResult.failed("Payment declined by issuer.");
         }
